goma: look at outputs when deriving exec root
`cargo` places the sources of dependencies into `CARGO_HOME` (which, for us, is
a subdirectory of the `cargo` workspace, e.g. `<workspace>/toolchain/cargo`) while
putting the compiled outputs into `<workspace>/target`. This causes goma to
(sometimes) use `<workspace>/toolchain` as common prefix for all artifacts (since all
inputs have that as prefix) and then fail the action because `ExpectedOutputDirs` does
not have that as prefix.

By looking at `ExpectedOutputFiles` and `ExpectedOutputDirs` when finding the
common prefix , we ensure that outputs are definitely in the exec root. This change
should not break any existing use-cases since we fail the action if an output is not in
the execution root. With this change, we instead fail to find an exec root if inputs and
outputs do not have any common prefix (or the prefix is `/`).

ENGFLOW_ORIGIN_REV_ID: e6fc0cc02ff3f327199b52f62036e02c577d77cf
Change-Id: If3d7c7d97dfcdd249a33aa7c46047e17c4f136df