goma: reset memory buffer when retrying download of std{out,err}
Previously, Goma would append to the already exiting buffer when
retrying a failing download (e.g. because of a download timeout),
which would result in duplicating the first `n` bytes of `std{out,err}`
(where `n` is less than the total size of the output). We've seen
this to cause failing actions on Windows where the includes of a
C++ file are printed on `std{our,err}` (and not in a `.d` file, as
it is on unix).

ENGFLOW_ORIGIN_REV_ID: b7fdd9877ec1b560000eff7be89fba460fbce53e
Change-Id: I1fcb6fd21ca34c8f9624bda9219cc3cf115d448f